requirements.txt
watchdog>=2.1.6
PyYAML>=5.4.1
cryptography>=3.4.7